Nutrient Show

+++ db/migrate/20201026195905_create_categories.rb
@@ -2,7 +2,7 @@ class CreateCategories < ActiveRecord::Migration[6.0]
   def change
     create_table :categories do |t|
       t.string :title
-      t.numeric :priority
+      t.integer :priority
 
       t.timestamps
     end
